backend/requirements.txt
asgiref==3.11.0
boto3==1.42.34
botocore==1.42.34
brotli==1.2.0
click==8.3.1
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.1
h11==0.16.0
jmespath==1.1.0
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
ruff==0.14.14
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
uvicorn==0.40.0
whitenoise==6.11.0
